[FIX] Cleanup
