+++ src/NHibernate.Caches.Redis/ILockingStrategy.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace NHibernate.Caches.Redis
+{
+    public interface ILockingStrategy
+    {
+        IDisposable ReadLock();
+
+        IDisposable WriteLock();
+    }
+}
