+++ src/NHibernate.Caches.Redis/ISerializer.cs
@@ -1,0 +1,9 @@
+namespace NHibernate.Caches.Redis
+{
+    public interface ISerializer
+    {
+        byte[] Serialize(object value);
+
+        object Deserialize(byte[] someBytes);
+    }
+}
